removed labs scenes, moved general purpose functions from starting to base scene, implemented frustum scene, fixed most warnings

+++ diplom_clean/ImGui/ImGuiLayer.h
@@ -15,7 +15,7 @@ public:
 	~ImGuiLayer();
 
 	void Begin();
-	void End(float width, float height);
+	void End(unsigned int width, unsigned int height);
 private:
 	std::string name;
 };
